backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pdfplumber>=0.10.3
python-docx>=1.1.0
spacy>=3.7.2
requests>=2.31.0
python-multipart>=0.0.7
aiofiles>=23.2.1
python-dotenv>=1.0.0
email-validator>=2.1.0
numpy>=1.24.0
